+++ src/components/Toggle.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { useColorMode, IconButton } from '@chakra-ui/react'
+import { MoonIcon, SunIcon } from '@chakra-ui/icons'
+
+export default function Toggle() {
+  const { colorMode, toggleColorMode } = useColorMode()
+  return (
+    <IconButton
+      aria-label="Toggle color mode"
+      icon={colorMode === 'light' ? <MoonIcon /> : <SunIcon />}
+      onClick={toggleColorMode}
+    />
+  )
+}
